refactor: moved holdings presenters into their own files
