+++ local-gems/spectrum-json/lib/spectrum/json.rb
@@ -96,7 +96,11 @@ require 'spectrum/request/profile'
 require 'spectrum/request/ids'
 require 'spectrum/request/debug'
 
+require 'spectrum/presenters'
 require 'spectrum/presenters/holding_presenter'
+require 'spectrum/presenters/holding_presenter/alma_holding_presenter'
+require 'spectrum/presenters/holding_presenter/elec_holding_presenter'
+require 'spectrum/presenters/holding_presenter/hathi_trust_holding_presenter'
 require 'spectrum/presenters/physical_item_presenter'
 require 'spectrum/presenters/electronic_item_presenter'
 require 'spectrum/holding/physical_item_status'
