+++ local-gems/spectrum-json/lib/spectrum/presenters.rb
@@ -1,0 +1,5 @@
+module Spectrum
+  module Presenters
+  end
+end
+
